Fin des listes : modification de listes

+++ sync_list.c
@@ -103,6 +103,22 @@ void supprimer_element_liste(LISTE* liste, int position)
 	}
 }
 
+void ajouter_nouveau_element_liste (LISTE* liste, char* nomFichier)
+{
+	FICHIER infos;
+	time_t temps;
+	char commande[TAILLE_MAX] = "touch ";
+	
+	strcat(commande, nomFichier);
+	system(commande);
+	
+	infos.nom = malloc(strlen(nomFichier) * sizeof(char));
+	strcpy(infos.nom,nomFichier);
+	temps = time(NULL);
+	infos.date = *localtime(&temps);
+	ajouter_element_liste(liste,infos);	
+}
+
 void lire_fichier (char* nomFichier, LISTE* liste)
 {
 	FILE* fichier = NULL;
@@ -120,7 +136,7 @@ void lire_fichier (char* nomFichier, LISTE* liste)
 		if(strcmp(infos.nom,nomFichier) != 0)
 		{
 			temps = time(NULL);
-			infos.date = localtime(&temps);
+			infos.date = *localtime(&temps);
 			ajouter_element_liste(liste,infos);
 		}		
 	}
@@ -128,23 +144,48 @@ void lire_fichier (char* nomFichier, LISTE* liste)
 	fclose(fichier);
 }
 
+void modifier_element_liste (LISTE* liste, char* nomElement)
+{
+	ELEMENT* curseur = liste->deb_liste;
+	time_t temps;
+	
+	while (strcmp(curseur->fichier.nom, nomElement))
+	{
+		curseur = curseur->suivant;
+		if (curseur == NULL)
+		{
+			printf("Le fichier n'a pas été trouvé");
+			return;
+		}
+	}
+	temps = time(NULL);
+	curseur->fichier.date = *localtime(&temps);
+
+}
+
 void afficher_liste(LISTE liste)
 {
 	ELEMENT* curseur = liste.deb_liste;
 	
+	printf("Début d'affichage\n");
+	
 	while (curseur != NULL)
 	{
-		printf("%s | %s\n",curseur->fichier.nom, asctime(curseur->fichier.date));
+		printf("%s | %s",curseur->fichier.nom, asctime(&curseur->fichier.date));
 		curseur = curseur->suivant;
 	}
+	
+	printf("\n");
 	
 }
 
 int main (void)
 {
+	pid_t PIDProd, PIDBack;	
 	LISTE* liste = creer_liste_vide();
 	
-	creer_liste_vide(liste);
+	system("touch fichier1.txt");
+	system("touch fichier2.txt");
 		
 	system("ls > out.txt");
 	lire_fichier("out.txt",liste);
@@ -152,8 +193,33 @@ int main (void)
 	
 	afficher_liste(*liste);
 	
-	while (liste->taille != 0)
-		supprimer_element_liste(liste,0);
+	PIDProd = fork();
+	
+	/*if (PIDProd != 0)
+		PIDBack = fork();*/ // Pas bseoin pour le moment
+		
+	
+	if (PIDProd == 0)
+	{
+		sleep(2);
+		modifier_element_liste (liste, "fichier1.txt");
+		sleep(3);
+		ajouter_nouveau_element_liste(liste, "fichier3.txt");
+		afficher_liste(*liste);
+	}
+	
+	else
+	{
+		wait(NULL);
+		afficher_liste(*liste);
+		
+		system("rm fichier1.txt");
+		system("rm fichier2.txt");
+		system("rm fichier3.txt");
+	
+		while (liste->taille != 0)
+			supprimer_element_liste(liste,0);
+	}
 	
 	return 0;
 }
